DjangoApi_project/requirements.txt
Package                       Version
----------------------------- -------
Django                        2.2.1  
djangorestframework           3.8.2  
djangorestframework-simplejwt 4.3.0  
Pillow                        6.0.0  
pip                           19.1.1 
pkg-resources                 0.0.0  
psycopg2                      2.8.2  
PyJWT                         1.7.1  
pytz                          2018.9 
setuptools                    41.0.1 
sqlparse                      0.3.0  
wheel                         0.33.4 
